+++ src/middleware/headers.js
@@ -4,6 +4,7 @@ var os = require('os');
 var winston = require('winston');
 
 var meta = require('../meta');
+var languages = require('../languages');
 
 module.exports = function (middleware) {
 	middleware.addHeaders = function (req, res, next) {
@@ -60,4 +61,32 @@ module.exports = function (middleware) {
 
 		next();
 	};
+
+	let langs = [];
+	middleware.autoLocale = function (req, res, next) {
+		if (parseInt(req.uid, 10) > 0 || !meta.config.autoDetectLang) {
+			return next();
+		}
+
+		var lang = req.acceptsLanguages(langs);
+		if (!lang) {
+			return next();
+		}
+		req.query.lang = lang;
+		next();
+	};
+
+	languages.listCodes(function (err, codes) {
+		if (err) {
+			winston.error('[middleware/autoLocale] Could not retrieve languages codes list!');
+			codes = [];
+		}
+
+		winston.verbose('[middleware/autoLocale] Retrieves languages list for middleware');
+		var defaultLang = meta.config.defaultLang || 'en-GB';
+
+		langs = [defaultLang].concat(codes).filter(function (el, i, arr) {
+			return arr.indexOf(el) === i;
+		});
+	});
 };
